fin-openai/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
tenacity==9.0.0
requests==2.32.3
httpx==0.27.2
azure-identity==1.17.1
azure-search-documents==11.6.0
azure-keyvault-secrets==4.9.0
openai==1.51.2
pytest==8.3.2
